multiple images functionality update
not yet completed

+++ report/models.py
@@ -5,7 +5,7 @@ from datetime import datetime
 class Report(models.Model):
     user = models.ForeignKey(User, on_delete=models.SET_NULL,null=True)
     description = models.TextField(null=True,blank=True)
-    floodLevel = models.IntegerField()
+    floodLevel = models.FloatField()
     latitude = models.FloatField()
     longitute = models.FloatField()
     timestamp = models.DateTimeField(default = datetime.now)
